Perform a fuzzy sort when sorting the top list.  Tracks within 5% of each other can now sometimes be considered to be in the wrong order.
A side effect of this is that if two tracks have the same score,
either one can be sorted before the other one, and which one wins
changes from time to time.

To humans, 55% and 56% are almost the same score.  With this change,
it is to mserv as well.  The closer two scores are, the larger the
chance that they are sorted backwards.  Scores further apart than 5%
are always sorted in the correct order.