Fix bug where minimum runtime was always zero.
Fix problems with line wrapping on starting play of new song.  If a
song with a long enough name was played, the string "resumed by root"
was split into "resumed" and "by root", which made Expect fail to
match "resumed by root".  Matching only "resumed" should be enough and
won't be wrapped.