If module_output_sync returns a value > 0, it now means the plugin isn't done with all output yet.  Unfinished modules are called again and again until they are done, before new input is fetched from the player app.
As MSERV_SUCCESS is 0, and MSERV_FAILURE is -1 (which is still interpreted
as an error), this change is backwards compatible with returning only
MSERV_SUCCESS / MSERV_FAILURE.

Also, the ossaudio plugin now uses O_NONBLOCKing writes to the sound card.
This way, instead of standing still waiting for the soundcard to finish all
output, mserv now reacts to user commands while waiting for output to
finish.

This brings the worst-case time of the play-stop cycle down from about
1200ms to about 200ms, and there is a huge improvement in the interactive
feel of mserv using the ossaudio plugin.