Use the play durations in the history when calculating satisfaction. The idea is that previously, satisfaction was updated per song, but now satisfaction is updated per millisecond.
For example, let's say somebody listened to two songs, one AWFUL and
one SUPERB.  The AWFUL one is one second, and the SUPERB one is five
minutes:

Song 1, AWFUL,  1 second
Song 2, SUPERB, 5 minutes (= 300 seconds)

With the previous scheme, after listening to both songs, the user's
satisfaction would be 50% / NEUTRAL, since the average of the two
songs is 50%.

With the new scheme, the user's satisfaction will be close to 100%,
since instead of "One AWFUL song and one SUPERB song", the
satisfaction is based on "One AWFUL second and 300 SUPERB seconds".

I think this is more fair than the previous scheme.