After a song has finished playing, compare the actual duration of the song with the one in our database.  If our database doesn't match reality, update the database.
If a song gets terminated prematurely (by "stop", "next" or whatever),
a similar comparision is made, but the database is updated only if the
song has been playing for longer than the database says it should.
That way, if the database says zero, and the song is interrupted after
ten seconds, the database is updated to say ten seconds instead.  That
still won't be correct, but it will be closer to the truth than the
previous value.