Handle the currently playing song better with respect to the satisfaction value now depending on how long each song in the history has played.
Now, when calculating the satisfaction, we assume that the currently
playing song will play to the end, and that we know for sure how long
it is.  This is the common case.

If the assumption is wrong, either because the song is terminated
prematurely or because our opinion about how long it is was wrong, the
history, people's satisfaction and the top list are all adjusted after
the song has ended.