Basic auto-factor implementation that should keep all users happy while still maintaining somewhat random music selection.
If a new user logs in, the factor is set to 0.99.

Then, each time a new song is put in the history, the user with the lowest
satisfaction rating of all logged in users is examined.  The factor is
raised or lowered depending on how that user's satisfaction compares to
a goal satisfaction value.

The goal is currently fixed at 0.75.