When calculating a user's satisfaction, count unrated songs as only half as long as properly rated songs.
The point is that if a song has been properly rated, we know for sure
what the user thinks of it.  However, if a song hasn't been properly
rated, we're just guessing at NEUTRAL.  With this change, our guesses
count only half as much as actual facts when calculating the
satisfaction.