Fairness improvements.
A user that hasn't heard any songs will now use the goal satisfaction as
satisfaction rather than 0.5.

The user's satisfaction is now squared when computing weights for the
different users.  This makes MServ more aggressive in trying to even out
differences between users' different tastes.

When a user's satisfaction is computed, the user's opinion of the song
playing at log in time is now taken into consideration.

Aim for 90% satisfaction (previously 99%) if only one person is logged in.