Improve boredom prevention algorithm.
Previously, if a user hadn't heard anything new for an hour, the factor
got set to 0.5.

Currently, if a user hasn't heard anything new for an hour, the users'
opinion about rated songs are halved.  Thus, when "bored", if a user has
rated a song SUPERB, that ratings counts as 0.5.  Unrated songs are
counted as usual, and may get values between 0.0 and 1.0.

In the new scheme, the factor isn't touched due to any boredom
considerations.